==> 4.Target Encoding.ipynb <==
import joblib
import pandas as pd
from sklearn.preprocessing import LabelEncoder

## Aligning Test Dataset Columns with Train Data

test_data = joblib.load('Encoded_Test_Data.joblib')

train_data = joblib.load('Encoded_Train_Data.joblib')

test_data.shape

train_data.shape

print(train_data.shape)
print(test_data.shape)

 **Test datasets are aligned in terms of their columns from Train datasets**

train_data, test_data = train_data.align(test_data, join='left', axis=1)
test_data.fillna(0, inplace=True)

test_data.shape

test_data = test_data.reset_index(drop=True)

print(train_data.shape)
print(test_data.shape)

## Target Encoding

a=test_data['IncidentGrade'].unique()
b=train_data['IncidentGrade'].unique()
a

b

## Unified Label Encoding: Combining, Fitting, and Transforming Labels in the Target Data

#combining the unique labels from both datasets
all_labels = list(set(train_data['IncidentGrade']).union(set(test_data['IncidentGrade'])))

#fiting the LabelEncoder on the combined labels
le = LabelEncoder()
le.fit(all_labels)

#transforming the labels
train_data['IncidentGrade'] = le.transform(train_data['IncidentGrade'])
test_data['IncidentGrade'] = le.transform(test_data['IncidentGrade'])

print("Train encoded labels:", train_data['IncidentGrade'].unique())
print("Test encoded labels:", test_data['IncidentGrade'].unique())
print("Label mapping:", dict(zip(le.classes_, le.transform(le.classes_))))


### Saving the Train & Test data using Joblib

joblib.dump(test_data, 'Encoded_Test_Data_F.joblib')

joblib.dump(train_data, 'Encoded_Train_Data_F.joblib')
